rbd: simplify rbd_merge_bvec()
When an rbd disk gets initialized, its block device queue gets
supplied with a merge_bvec function.  This is a callback from the
block queue code used to allow the device to indicate whether an
existing bio_vec can be extended to include another request
contiguous in memory.

The interface this merge function uses is a little strange, but the
bottom line is the caller is asking whether adding a new bio_vec to
an existing bio produces an acceptable result.  The way the result
is used means it can be interpreted as asking, "to what length can
I extend this request before I hit a device-defined boundary?"
I.e., the caller always tests whether the merge result is less
than the length of the proposed addition, and if it is, doesn't
add it.


The merge function supplied by rbd is rbd_merge_bvec(), and it is
defined in a sort of convoluted way.  (This is the case for most of
the other merge functions, by the way.)

It is also incorrect, and produces a result that's inconsistent with 
what rbd_get_segment() returns as its segment length.

What's incorrect is that it's adding the size of the *existing bio*
in computing whether the merge should be allowed, when it should
be adding the size of the *proposed addition to the bio.*

What that means is that the function precludes many merges from
occuring that should be allowed, forcing something like half of
consecutive block requests to be forced to go one chunk at a time.


In any case, rather than go into any more detail...  This patch
dramatically simplifies this merge function, simply returning how
much space (in bytes) is left the rbd image segment that contains
the starting offset for the given I/O.

Use "seg" rather than "chunk" to indicate the units whose boundaries
we care about we call "segments."

It also adds an extra validation check to rbd_dev_ondisk_valid(),
ensuring the order is at least as big as a sector (a requirement
of the generic bio code).

Signed-off-by: Alex Elder <elder@inktank.com>