ceph: always set inode size with exclusive caps
If the exclusive caps are held for a file, on a truncate
set the i_size and avoid sending the setattr request to the mds.
Also prepare the truncate fields for __ceph_do_pending_vmtruncate.

This fixes http://tracker.newdream.net/issues/3681, where the i_size
was only being set if it was smaller than the requested size, otherwise
a setattr request was sent to the mds, which could race with a cap
update coming from a write completion.

Signed-off-by: Sam Lang <sam.lang@inktank.com>