ceph: set inode i_size before truncate request
On a truncate down set the i_size before sending the setattr
request to the mds to ensure that cap flushes also send the
correct (truncated) size to the mds.

This fixes http://tracker.newdream.net/issues/3681, where a
setattr request was sent to the mds without first setting the
i_size, which could race with a cap update coming from
write completion.

Signed-off-by: Sam Lang <sam.lang@inktank.com>